restructure anvio.workflows folder
